pyproject.toml
[project]
name = "summer-project-app"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "crewai>=0.152.0",
    "crewai-tools>=0.59.0",
    "dspy>=2.6.13",
    "exa-py>=1.14.20",
    "gitpython>=3.1.45",
    "google-generativeai>=0.8.5",
    "pygithub>=2.7.0",
    "python-dotenv>=1.1.1",
]
